pyproject.toml
[project]
name = "b2broker-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Lev Borodin",email = "faunris@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (==5.1.7)",
    "django-filter (==25.1)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-jsonapi (==7.1.0)",
    "psycopg[binary] (==3.2.6)",
    "psycopg-pool (==3.2.6)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
factory-boy = "3.3.3"
pytest = "8.3.5"
pytest-django = "4.10.0"
pytest-factoryboy = "2.7.0"
pre-commit = "4.1.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "app.settings"

[tool.ruff]
extend-exclude = ["*/migrations"]
line-length = 119
